# e-kavach-backend/openapi.yaml
openapi: 3.0.3
info:
  title: E-KAVACH Emergency Healthcare & Clinical Network API
  version: 1.0.0
  description: |
    Production-grade backend service powering E-KAVACH emergency healthcare network,
    ABHA citizen integration, Golden Hour emergency triage scanner, and hospital bed telemetry.
servers:
  - url: http://localhost:5000/api
    description: Local Backend Service
  - url: https://api.ekavach.health/api
    description: Production Gateway
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: ekavach_access_token
...
paths:
...
  /patient/me:
    get:
      summary: Retrieve authenticated patient profile
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: Patient profile
  /patient/abha:
    get:
      summary: Get linked ABHA number and ABDM profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: ABHA details
  /patient/emergency-pass:
    get:
      summary: Get active Golden Hour emergency medical pass
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Emergency pass details
...
  /patient/records:
    post:
      summary: Upload medical lab report or prescription
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                document:
                  type: string
                  format: binary
                title:
                  type: string
                recordType:
                  type: string
      responses:
        '201':
          description: Record uploaded
...